DeepSeekChat/obj/requirements.txt
flask
python-dotenv
requests
python-multipart
PyPDF2       # For PDF processing (optional)
python-docx  # For Word docs (optional)